refac: rearrange files
