+++ .gitignore
@@ -5,7 +5,6 @@ Cargo.lock
 bin/
 pkg/
 wasm-pack.log
-worker/
 node_modules/
 .cargo-ok
