+++ frontend/utils/utils.ts
@@ -1,4 +1,5 @@
-import { NAME_REGEX, parseExpiration, parseExpirationReadable, PASSWD_SEP } from "../../src/shared.js"
+import { NAME_REGEX, PASSWD_SEP } from "../../shared/constants.js"
+import { parseExpiration, parseExpirationReadable } from "../../shared/parsers.js"
 
 export const BaseUrl = DEPLOY_URL
 export const APIUrl = API_URL
