+++ shared/parsers.ts
@@ -1,40 +1,4 @@
-// This file contains things shared with frontend
-
-export type PasteLocation = "KV" | "R2"
-
-export type PasteResponse = {
-  url: string
-  suggestedUrl?: string
-  manageUrl: string
-  expirationSeconds: number
-  expireAt: string
-}
-
-export type MetaResponse = {
-  lastModifiedAt: string
-  createdAt: string
-  expireAt: string
-  sizeBytes: number
-  location: PasteLocation
-  filename?: string
-  encryptionScheme?: string
-}
-
-export type MPUCreateResponse = {
-  name: string
-  key: string
-  uploadId: string
-}
-
-export const CHAR_GEN = "ABCDEFGHJKMNPQRSTWXYZabcdefhijkmnprstwxyz2345678"
-export const NAME_REGEX = /^[a-zA-Z0-9+_\-[\]*$@,;]{3,}$/
-export const PASTE_NAME_LEN = 4
-export const PRIVATE_PASTE_NAME_LEN = 24
-export const DEFAULT_PASSWD_LEN = 24
-export const MAX_PASSWD_LEN = 128
-export const MIN_PASSWD_LEN = 8
-export const MAX_URL_REDIRECT_LEN = 2000
-export const PASSWD_SEP = ":"
+import { PASSWD_SEP } from "./constants.js"
 
 export function parseSize(sizeStr: string): number | null {
   sizeStr = sizeStr.trim()
