+++ worker/handlers/handleCors.ts
