+++ worker/handlers/handleMPU.ts
@@ -1,4 +1,5 @@
-import { MPUCreateResponse, NAME_REGEX, PASTE_NAME_LEN, PRIVATE_PASTE_NAME_LEN } from "../shared.js"
+import { MPUCreateResponse } from "../../shared/interfaces.js"
+import { NAME_REGEX, PASTE_NAME_LEN, PRIVATE_PASTE_NAME_LEN } from "../../shared/constants.js"
 import { genRandStr, WorkerError } from "../common.js"
 import { getPasteMetadata, pasteNameAvailable } from "../storage/storage.js"
