+++ worker/test/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "types": ["@cloudflare/vitest-pool-workers", "../../worker-configuration.d.ts"],
+    "moduleResolution": "bundler",
+    "strict": true
+  },
+  "include": ["**/*.ts"]
+}
